Change GitBlameLine.withCValue(_:) from rethrows to throw
`rethrows` is technically valid and compiles without errors, since
`throw NSError.makeCConversionError()` occurs within the
`withUnsafeBytes(_:)` closure, which is itself a `rethrows` method.

However, the Swift documentation says that `rethrows` is intended to
"indicate that [a function] throws an error only if one of its function
parameters throws an error."

`rethrows` is semantically misleading in this case, since the method
can throw an error independent of the `body` parameter.