Update `String` parameters to `Data`
`String` may be used as a binding type only when the following conditions
are met:
- The libgit2 documentation explicitly describes the field as a
null-terminated string.
- The libgit2 documentation does not mention unknown encoding, bytes,
or binary/blob/diff/raw data.
- There is no associated length field.

Otherwise, `Data` should be used to safely represent the complete C value.

If the `baseAddress` of the bytes returned by `withUnsafeBytes(_:)` is
`nil`, bindings treat this as an error. This catches edge cases where
`baseAddress` might be `nil` for non-empty data, and avoids undefined
behavior as a result of passing null buffers to C functions that may
not validate the buffers at runtime.