Use literal macro values for `GitAttrCheckFlagT` options
The options of `GitAttrCheckFlagT` are represented in libgit2 as
macros. They are represented in swift-libgit2 as an enum for
consistency with other flags and for type-safe usage.

C preprocessor macros have no inherent type. Most are represented in
Swift as `Int32` as an artifact of the Swift-C bridging layer, but are
generally used throughout the API as `UInt32` or other numeric types.
Hard-coding literal values with the correct Swift type eliminates
potentially unsafe type casting.

The `gitAttrCheckFileThenIndex` option uses `UInt32(0)` instead of `0`
for consistency with other options, even though the cast is not
necessary. This silences a compiler warning about creating an empty
option set with a literal zero value.