Make `GitAttrCheckFlagsT` conform to `GitOptionSet`
Previously, `GitAttrCheckFlagsT` conformed to only `OptionSet` and
`Sendable`. Unlike other option sets, this struct has no equivalent
C type since the flags are defined as macros. However, by implementing
`init(cValue:)` and `cValue()` using the raw value directly, it can now
conform to `GitOptionSet` for consistency across the API and to ensure
that it is not overlooked when protocol requirements change.