Return `GitErrorCode` raw values from callback tests
Callback tests now return `GitErrorCode` raw values instead of
integer literals, for improved semantics and maintainability.

The callbacks also return `gitUnknown(-123)` if the payload is `nil`.
This avoids overlap with any error code libgit2 may return, and also
avoids overlap with `gitEUser`, which may be returned if a Swift-to-C
conversion error occurred during the test.