Replace `GitBuf` with `Data`
Change `GitBuf` to a documentation-only struct and instead use `Data`
for all output buffer parameters. `Data` is more idiomatic Swift and
accurately represents the string, binary, or raw data that libgit2
returns through `git_buf`. This also relieves callers of the
responsibility of calling `gitBufDispose(buffer:)` with the `GitBuf`
instance when it is no longer needed.