Remove null terminator from `git_strarray` in `withGitStrArray(_:)`
`withArrayOfCStrings(_:)` appends a null terminator as the last element
of the array, but `git_strarray` must not include this terminator,
otherwise it will cause a runtime crash when libgit2 tries to read
invalid memory.

Also, the method now returns an `UnsafePointer` instead of a mutable
pointer, since libgit2 only reads from the `git_strarray`.