Standardize default struct option set values
Only some option sets have explicit default values that are equivalent
to an empty option set. For consistency, use an empty option set for
the default value of any `COptionSet` struct property.

For the option sets with explicit default values, tests still assert
that the default value is equal to the named default option rather than
an empty option set.