Remove `GitStashSaveOptions.init(cValue:)` warning
The warning is applicable to `GitBlameHunk` since that struct does not
have an associated initialization function. The warning is not
specifically applicable to this struct and others like it, since any
struct with an initialization function must always be properly
initialized before being used, or it will cause undefined behavior.