Remove DocC Discussion headers
The explicit Discussion headers work as expected with the generated
DocC documentation. However, they cause duplicate headers when viewing
Quick Help documentation, if the API is a function that has a `Throws`
line in its documentation. For consistency, remove the headers from all
APIs, not just from functions, and use the generated headers.